requirements.txt
pandas==0.20.3
requests==2.18.4
scikit-learn==0.19.2
scipy==0.19.1
Flask-Babel==0.9
